# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApnaFarm API",
        "version": "1.0.1"
    },
    "paths": {
...
        "/api/add-farm-details": {
            "post": {
                "tags": [
                    "Farm"
                ],
                "summary": "Create a new farm",
                "description": "Creates a new farm with location, size, configuration details, and crop stage.",
                "operationId": "21361b7e13669ecf0f55ca8b256ceeb2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "location",
                                    "size",
                                    "latitude",
                                    "longitude",
                                    "irrigation_source",
                                    "soil_type",
                                    "sowing_method",
                                    "seed_variety",
                                    "crop",
                                    "sowing_date",
                                    "farm_configuration",
                                    "crop_stage"
                                ],
                                "properties": {
                                    "location": {
                                        "type": "string",
                                        "example": "Kampala"
                                    },
                                    "size": {
                                        "type": "string",
                                        "example": "10"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "0.3131"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "32.5811"
                                    },
                                    "irrigation_source": {
                                        "description": "Dropdown selection: well, canal, rainfed",
                                        "type": "string",
                                        "enum": [
                                            "well",
                                            "canal",
                                            "rainfed"
                                        ],
                                        "example": "well"
                                    },
                                    "soil_type": {
                                        "description": "Dropdown selection: clay, loam, sandy",
                                        "type": "string",
                                        "enum": [
                                            "clay",
                                            "loam",
                                            "sandy"
                                        ],
                                        "example": "clay"
                                    },
                                    "sowing_method": {
                                        "description": "Dropdown selection: manual, mechanized, hydroponic",
                                        "type": "string",
                                        "enum": [
                                            "manual",
                                            "mechanized",
                                            "hydroponic"
                                        ],
                                        "example": "manual"
                                    },
                                    "seed_variety": {
                                        "description": "Dropdown selection: hybrid, organic, GMO",
                                        "type": "string",
                                        "enum": [
                                            "hybrid",
                                            "organic",
                                            "GMO"
                                        ],
                                        "example": "hybrid"
                                    },
                                    "crop": {
                                        "description": "Dropdown selection: wheat, rice, maize",
                                        "type": "string",
                                        "enum": [
                                            "wheat",
                                            "rice",
                                            "maize"
                                        ],
                                        "example": "wheat"
                                    },
                                    "sowing_date": {
                                        "type": "string",
                                        "example": "2021-09-01"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Farm 1"
                                    },
                                    "farm_configuration": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "latitude": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "example": 10
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "crop_stage": {
                                        "description": "Dropdown selection: sapling_stage, vegetative_growth, flowering_stage, fruit_setting, fruit_development, harvesting, post_harvest_ripening",
                                        "type": "string",
                                        "enum": [
                                            "sapling_stage",
                                            "vegetative_growth",
                                            "flowering_stage",
                                            "fruit_setting",
                                            "fruit_development",
                                            "harvesting",
                                            "post_harvest_ripening"
                                        ],
                                        "example": "vegetative_growth"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Farm created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Farm created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/get-farm-list": {
            "get": {
                "tags": [
                    "Farm"
                ],
                "summary": "Get list of farms for the authenticated user",
                "operationId": "getFarmList",
                "responses": {
                    "200": {
                        "description": "Farm details retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/farm/enums": {
            "get": {
                "tags": [
                    "Farm"
                ],
                "summary": "Get available enums for farm",
                "description": "Display available enum options.",
                "operationId": "b742a6b8ae4d6d56a9ef6e06cc8b0660",
                "responses": {
                    "200": {
                        "description": "List of Enums",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "irrigation_source": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "soil_type": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "sowing_method": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "seed_variety": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "crop": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/farm-statistics": {
            "get": {
                "tags": [
                    "Farm"
                ],
                "summary": "Get farm statistics",
                "description": "This endpoint returns the total number of farms, total land size, and crop distribution (in acres).",
                "operationId": "a146f18540f7d0a40884d5860f0aa8c1",
                "responses": {
                    "200": {
                        "description": "Farm statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_farms": {
                                            "description": "The total number of farms",
                                            "type": "integer",
                                            "example": 29
                                        },
                                        "total_land": {
                                            "description": "The total size of land in acres",
                                            "type": "integer",
                                            "example": 120
                                        },
                                        "crops": {
                                            "description": "Breakdown of crop distribution in acres",
                                            "properties": {
                                                "Sugarcane": {
                                                    "description": "Acres used for Sugarcane",
                                                    "type": "integer",
                                                    "example": 29
                                                },
                                                "Cotton": {
                                                    "description": "Acres used for Cotton",
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "Wheat": {
                                                    "description": "Acres used for Wheat",
                                                    "type": "integer",
                                                    "example": 11
                                                },
                                                "Free": {
                                                    "description": "Acres of free land",
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "Enter your Bearer token in the format **Bearer &lt;token&gt;**",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "Bearer": []
        }
    ]
}